Fixed bug in call to `fits_read_colnull'.
As I discovered, `fits_read_colnull' does not accept a NULL pointer for
the `anynull' flag. Since this is accessed only when there are NULLs in
the data read by the routine (i.e. the function does not set it to 0
when no NULLs are found), the bug appeared only when reading columns
which actually contained NULL values.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rfitsio@93 edb9625f-4e0d-4859-8d74-9fd3b1da38cb